+++ test/fixture/src/ReactJSX.js
@@ -1,0 +1,11 @@
+/**
+ * @private
+ */
+export default class ReactJSXClass {
+  /**
+   * this method use React JSX syntax.
+   */
+  render() {
+    return <div>Hello</div>;
+  }
+}