+++ test/src/BuilderTest/CoverageDocTest.js
@@ -8,9 +8,9 @@ describe('Coverage:', ()=> {
   it('has coverage.json', ()=>{
     let json = fs.readFileSync('./test/fixture/esdoc/coverage.json', {encoding: 'utf8'}).toString();
     let coverage = JSON.parse(json);
-    assert.equal(coverage.coverage, '81.01%');
-    assert.equal(coverage.expectCount, 79);
-    assert.equal(coverage.actualCount, 64);
+    assert.equal(coverage.coverage, '81.48%');
+    assert.equal(coverage.expectCount, 81);
+    assert.equal(coverage.actualCount, 66);
     assert.deepEqual(coverage.files, {
       "src/ForTestDoc/AbstractDoc.js": {
         "expectCount": 3,
@@ -43,6 +43,10 @@ describe('Coverage:', ()=> {
       "src/OtherClass/SuperMyClass.js": {
         "expectCount": 19,
         "actualCount": 19
+      },
+      "src/ReactJSX.js": {
+        "actualCount": 2,
+        "expectCount": 2
       },
       "src/myFunction.js": {
         "expectCount": 8,